Create Max_Consecutive_Ones.js
# My First Leetcode Solution. 

I have not taken Data Structures and Algorithms yet, and so far have only learned the basic syntax of Javascript and other programming languages such as Python. None of my introductory programming courses have any focus on solving coding problems. 

But, I really would love for the opportunity to get an internship before I graduate from WGU. Since it's a non-traditional college, it's hard to gauge the equivalent "year" of study I'm in with accelerated courses, so I thought I would just start applying now. The problem is that, through my research, most of these internships interviews still do want students who have already taken DSA, and therefore tend to have coding assessments using LC or Hackerrank. I looked at my program schedule though, and my DSA class isn't for several terms - months away if I continue accelerating, perhaps years if I start only taking the standard amount of classes per semester. I don't want to wait that long to get an internship when I could just be practicing on my own. 

So I went to Leetcode Explore and lo and behold, they have a sort of introductory course on solving coding problems. So that's where I found this one. 

## Thought Process 

To start this problem, I first went through and wrote in plain English how I would have to solve the problem in coding terms. Some of that write up remains in the comments, but not all of it. The goal of doing this was to break down the problem into specific steps I knew could be represented by a bit of code, so that even if I couldn't figure out how to write the code for it, I would at least get the practice of how to think about coding problems if nothing else. I went into this not really expecting to be able to solve it at all. I've scrolled through Leetcode easy's and Codewars 7kyu's hundreds of times and even the simplest problems seemed to be beyond me. At this point, going through this tutorial was sort of a last resort thing where I thought to myself, "If I can't get it with this tutorial, maybe I really do just need to wait until I take the class..." 

But I could still at least try. So I did. 

I had to look up the loop syntax again... every time I write a for loop I seem to have to do that. For some reason I still can't remember it off the top of my head at this point. On the first test case I ran, I just wanted to check if the loop was working, I got an error for an "Unexpected Token"... and realized I'd used a comma after nums.length instead of a semicolon. So I'm still making very elementary mistakes with syntax, but at least I was able to fix it. 

I'm pretty sure this solution is the absolute furthest from elegant it could be, but I was really proud of myself that it is the very first coding problem I have ever solved on my own without having to google a solution and try to understand what someone else did and replicate it. On this one, the only thing I had to look up was the loop syntax. Everything else I figured out for myself. 

For the coding champions and rock stars, this probably seems like the simplest, dumbest thing to be proud of, but for someone who didn't even start coding until 2 months ago, this is still an accomplishment for me! I'm happy to have successfully completed ANY problem on my own, even if it's the most rudimentary one LC could throw at me - because just yesterday learning how to do this on my own seemed insurmountable, and now it seems maybe that goal isn't as unreachable as I thought it was. 

I look forward to going through the rest of LC's tutorials. Maybe what I needed was just the right kind of explanation. :)